requirements.txt
panda
matplotlib
numpy
seaborn
skimpy
scikit-learn
alembic
sqlalchemy
pydantic
fastapi
pytest
joblib
loguru
mlflow
optuna
lightgbm
streamlit